refactor(client): separate http, websocket & grpc
